+++ src/pymordial/core/__init__.py
@@ -4,7 +4,6 @@ Core module for Pymordial.
 
 from pymordial.core.app import PymordialApp
 from pymordial.core.blueprints.bridge_device import PymordialBridgeDevice
-from pymordial.core.blueprints.element import PymordialElement
 from pymordial.core.blueprints.emulator_device import (
     EmulatorState,
     PymordialEmulatorDevice,
@@ -13,6 +12,7 @@ from pymordial.core.blueprints.vision_device import PymordialVisionDevice
 from pymordial.core.controller import PymordialController
 from pymordial.core.screen import PymordialScreen
 from pymordial.core.state_machine import AppState, StateMachine
+from pymordial.ui.element import PymordialElement
 from pymordial.ui.image import PymordialImage
 from pymordial.ui.pixel import PymordialPixel
 from pymordial.ui.text import PymordialText